New Feature: Support for "KernelVariables": SYSFS like filesystem efficiently exporting arbitrary variables from kernel space (currently only RO) - Bugfixes for race conditions and memory management bugs    -> This version seams to be very stable for now. Ran multiple tests, in field system is now running 5 days with 8 tasks and 15 rt threads with periods between 1ms - 500 ms doing I/O as well as aperiodic threads doing network (TCP/IP, UDP transfer).         Highest task entry latency measured 7 us (further investigations ongoing to drop this to 2 us) on BeagleBoardXM with 600 mhz - moved some architecture indepenent classes to shared arch directory - FAT file deletion support (beware: still experimental) - thread arrival time now given as absolute 64 bit cycle time! - telnet server enhanced:    - support for ps, ls -lh and df commands using the sysfs (kernelvariable) architecture

+++ source/comm/FixedSizePBufList.h
@@ -24,7 +24,7 @@ class FixedSizePBufList
     size_t size;
 
 public:
-    FixedSizePBufList(int size);
+    FixedSizePBufList(size_t size);
 
     ~FixedSizePBufList();
